Added makefile and moved header files to include/
