+++ .gitignore
@@ -1,4 +1,4 @@
-a.out
-a.out.DSYM
+build/
+*.DSYM
 test
 .DS_Store
